Rename Toast component as ToastPresenter
Intending to make the default export a ToastPresenter wrapped in a
react-transition
